Remove upload of media if is GIPHY

+++ victorious/victorious/Operations/Upload Operations/CreateMediaUploadOperation.swift
@@ -34,11 +34,13 @@ class CreateMediaUploadOperation: BackgroundOperation {
         upload(uploadManager)
     }
     
-    private func upload(uploadManager: VUploadManager) {
-        guard let mediaURL = formFields["media_data"] as? NSURL where !mediaURL.absoluteString.isEmpty else {
-            uploadCompletion(NSError(domain: "UploadError", code: -1, userInfo: nil))
-            return
+    private func completionError() {
+        dispatch_async(dispatch_get_main_queue()) {
+            self.uploadCompletion(NSError(domain: "UploadError", code: -1, userInfo: nil))
         }
+    }
+    
+    private func upload(uploadManager: VUploadManager) {
         let taskCreator = VUploadTaskCreator(uploadManager: uploadManager)
         taskCreator.request = request.urlRequest
         taskCreator.formFields = formFields
@@ -48,11 +50,13 @@ class CreateMediaUploadOperation: BackgroundOperation {
             let task = try taskCreator.createUploadTask()
             uploadManager.enqueueUploadTask(task) { _ in }
         } catch {
-            uploadCompletion(NSError(domain: "UploadError", code: -1, userInfo: nil))
+            completionError()
             return
         }
         
-        let _ = try? NSFileManager.defaultManager().removeItemAtURL(mediaURL)
+        if let mediaURL = formFields["media_data"] as? NSURL {
+            let _ = try? NSFileManager.defaultManager().removeItemAtURL(mediaURL)
+        }
         
         dispatch_async(dispatch_get_main_queue()) {
             self.uploadCompletion(nil)
@@ -63,11 +67,13 @@ class CreateMediaUploadOperation: BackgroundOperation {
     private var formFields: [NSObject : AnyObject] {
         var dict: [NSObject : AnyObject] = [
             "name" : publishParameters.caption,
-            "media_data" : mediaURL ?? NSURL(string: "")!,
             "is_gif_style" : publishParameters.isGIF ? "true" : "false",
             "did_crop" : publishParameters.didCrop ? "true" : "false",
             "did_trim" : publishParameters.didTrim ? "true" : "false",
         ]
+        if !publishParameters.isGIF {
+            dict["media_data"] = mediaURL ?? NSURL(string: "")!
+        }
         
         if let filterName = publishParameters.filterName {
             dict["filter_name"] = filterName
